Fix postRunProcessing behavior during docker build
If Windows updates are being applied via a RUN command in a docker build, the build will not stop if there was a failure in postRunProcessing.  To enable this behavior, we explicitly set the exit code of the container to a failure if postRunProcessing fails during servicing container step.  For completeness, also avoid running servicing operation if the original exit code of the container is non-zero so that original failure exit code does not get overwritten.

Signed-off-by: Stefan J. Wernli <swernli@microsoft.com>