Add --json to many of the CLI commands
Add a --json flag to most of the CLI commands, to force output that
would have gone to stdout to be output as JSON instead of whatever form
it would otherwise take.  Error messages sent to stderr are still plain
text.

Use this chance to rework the wrapping logic so that it's more
consistent: commands that take multiple IDs attempt the specified
operation on all of them, and report errors afterward.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>