Don't list container layers as ordinary ones
When computing dependency lists, don't list container layers in the
layers lists, since they're already going to be listed as containers,
and listing them twice will confuse callers who want to know what else
they need to delete when attempting to remove something.

This also assumes that no layers depend on container layers, so add a
check in CreateLayer to make sure we don't allow that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>