Require ImageDestination.PutBlob to fail and delete data on error in source stream
This is necessary to prevent using layers which have been tampered with.

To implement this, dir: and oci: now store data into temporary files,
and Rename() when all data is successfully copied and validated.

We do not do anything extra for docker:, hoping that the mandatory
digest= parameter in the API call is sufficient.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>