Add types.Image.EmbeddedDockerReferenceConflicts
This allows us to detect whether we need to edit the reference before
trying to (i.e. before uploading any of the layers).  The method
interface may be a bit surprising, it is designed to hide the
complexity and schema1 specifics in manifestSchema1; see the comment
in that implementation for more details.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>