Switch MakeStore() to expecting an options struct
Switch MakeStore() to accepting an options structure instead of
individual arguments, so that adding options in the future won't require
rewriting logic that calls it, per discussion in issue #2.

This also lets the CLI wrapper pick up its defaults from the library, so
it no longer has to have them hard-coded.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>