Revert "Unit test for golang gzip backward compatibility"
This reverts commit 6dfec68c4b3f55ac7b7ffab555525c65b0f02950.

Already golang 1.6.3 (e.g. Fedora 24) and 1.7 (the test values) differ
in gzip output, which makes development on Fedora 24 cumbersome.

To fix the gzip output, we would need to carry a forked version of the
gzip compression implementation; looking at
https://go.googlesource.com/go/+log/master/src/compress , the idea of
following upstream changes and deciding which ones are bugfixes to
backport and which ones are improvements to leave out is extremely
unattractive.

And of course bugfixes which need backporting may change
the output anyway; we simply can’t make a long-term promise that the
compressed stream will stay unchanged with an evolving implementation.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>