Track the last writer in lockfile.Modified()
Instead of checking if the last writer was us in the lockfile Modified()
method, check if the last writer (which changes every time the lockfile
is Touch()ed) changed since the last time we read it.  This should
produce the same results for a frequent writer, but keeps a
mostly-read-only participant from thinking that the data protected by
the lock has been changed every single time it checks, after someone
else has written it even once.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>